Split core types into a separate crate
They're useful even if you don't want to use Tracer etc
