+++ zipkin-types/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "zipkin-types"
+version = "0.1.0"
+authors = ["Steven Fackler <sfackler@palantir.com>"]
+license = "Apache-2.0"
+description = "Type definitions for Zipkin distributed trace information"
+repository = "https://github.com/palantir/rust-zipkin"
+readme = "../README.md"
+categories = ["network-programming", "web-programming"]
+keywords = ["zipkin", "tracing"]
+
+[dependencies]
+data-encoding = "2.1"
+serde = { version = "1.0", optional = true, features = ["derive"] }
