+++ zipkin-types/src/span_id.rs
