+++ zipkin-types/src/trace_id.rs
