+++ zipkin/src/lib.rs
@@ -27,97 +27,28 @@
 #![doc(html_root_url = "https://docs.rs/zipkin/0.3")]
 #![warn(missing_docs)]
 
-extern crate data_encoding;
 extern crate rand;
 extern crate thread_local_object;
+extern crate zipkin_types;
 
 #[macro_use]
 extern crate log;
 
-#[macro_use]
-#[cfg(feature = "serde")]
-extern crate serde;
-
-#[cfg(feature = "serde")]
-use serde::{Serialize, Serializer};
-#[cfg(feature = "serde")]
-use std::time::{Duration, SystemTime, UNIX_EPOCH};
-
 #[doc(inline)]
-pub use annotation::Annotation;
-#[doc(inline)]
-pub use endpoint::Endpoint;
+pub use zipkin_types::{annotation, endpoint, sampling_flags, span, span_id, trace_context,
+                       trace_id, Annotation, Endpoint, Kind, SamplingFlags, Span, SpanId,
+                       TraceContext, TraceId};
+
 #[doc(inline)]
 pub use report::Report;
 #[doc(inline)]
 pub use sample::Sample;
 #[doc(inline)]
-pub use sampling_flags::SamplingFlags;
-#[doc(inline)]
-pub use span::{Kind, Span};
-#[doc(inline)]
-pub use span_id::SpanId;
-#[doc(inline)]
-pub use trace_context::TraceContext;
-#[doc(inline)]
-pub use trace_id::TraceId;
-#[doc(inline)]
 pub use tracer::{OpenSpan, Tracer};
 
-pub mod annotation;
-pub mod endpoint;
 pub mod report;
 pub mod sample;
-pub mod sampling_flags;
-pub mod span;
-pub mod span_id;
-pub mod trace_context;
-pub mod trace_id;
 pub mod tracer;
 
 #[cfg(test)]
 mod test;
-
-#[cfg(feature = "serde")]
-fn time_micros<S>(time: &SystemTime, s: S) -> Result<S::Ok, S::Error>
-where
-    S: Serializer,
-{
-    duration_micros(
-        &time.duration_since(UNIX_EPOCH)
-            .unwrap_or(Duration::from_secs(0)),
-        s,
-    )
-}
-
-#[cfg(feature = "serde")]
-fn duration_micros<S>(duration: &Duration, s: S) -> Result<S::Ok, S::Error>
-where
-    S: Serializer,
-{
-    let micros = duration.as_secs() * 1_000_000 + duration.subsec_nanos() as u64 / 1_000;
-    let micros = micros.max(1);
-    micros.serialize(s)
-}
-
-#[cfg(feature = "serde")]
-fn opt_time_micros<S>(time: &Option<SystemTime>, s: S) -> Result<S::Ok, S::Error>
-where
-    S: Serializer,
-{
-    match *time {
-        Some(ref time) => time_micros(time, s),
-        None => s.serialize_none(),
-    }
-}
-
-#[cfg(feature = "serde")]
-fn opt_duration_micros<S>(duration: &Option<Duration>, s: S) -> Result<S::Ok, S::Error>
-where
-    S: Serializer,
-{
-    match *duration {
-        Some(ref duration) => duration_micros(duration, s),
-        None => s.serialize_none(),
-    }
-}
